requirements.txt
Flask
yt-dlp
moviepy
openai-whisper
torch
